Refactor position calculation to handle stock splits
This commit refactors the `Ledger.Position` method to accurately calculate security positions by taking stock splits into account. The position calculation now considers the split-adjusted quantity of a security on a given date.

The key changes are:
- The `Ledger.Position` method now requires a `MarketData` object to access split information.
- The position calculation logic in `Ledger.Position` has been updated to adjust the quantity of a security based on any splits that have occurred.
- The `eodhd.go` file has been modified to support fetching both adjusted and unadjusted prices.
- All calls to `Ledger.Position` have been updated to pass the `MarketData` object.

This change is a crucial step towards correctly handling stock splits throughout the portfolio management system.

Fixes #43

+++ cmd/update_security.go
@@ -15,8 +15,8 @@ type updateSecurityCmd struct {
 	id    string
 	date  string
 	price float64
-	num   int
-	den   int
+	num   int64
+	den   int64
 }
 
 func (*updateSecurityCmd) Name() string { return "update-security" }
@@ -39,8 +39,8 @@ func (c *updateSecurityCmd) SetFlags(f *flag.FlagSet) {
 	f.StringVar(&c.id, "id", "", "Unique security identifier (required)")
 	f.StringVar(&c.date, "d", date.Today().String(), "Date for the update")
 	f.Float64Var(&c.price, "p", 0, "Price to set for the security")
-	f.IntVar(&c.num, "num", 1, "Split numerator")
-	f.IntVar(&c.den, "den", 1, "Split denominator")
+	f.Int64Var(&c.num, "num", 1, "Split numerator")
+	f.Int64Var(&c.den, "den", 1, "Split denominator")
 }
 
 func (c *updateSecurityCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}) subcommands.ExitStatus {
